+++ api/v1/server/authz/middleware.go
@@ -3,28 +3,24 @@ package authz
 import (
 	"fmt"
 	"net/http"
-	"os"
-	"path/filepath"
-	"runtime"
-	"strings"
 
 	"github.com/labstack/echo/v4"
 	"github.com/rs/zerolog"
-	"go.yaml.in/yaml/v4"
 
 	"github.com/hatchet-dev/hatchet/api/v1/server/middleware"
+	"github.com/hatchet-dev/hatchet/api/v1/server/rbac"
 	"github.com/hatchet-dev/hatchet/pkg/config/server"
 	"github.com/hatchet-dev/hatchet/pkg/repository/sqlcv1"
 )
 
 type AuthZ struct {
 	config *server.ServerConfig
-	rbac   *Authorizer
+	rbac   *rbac.Authorizer
 	l      *zerolog.Logger
 }
 
 func NewAuthZ(config *server.ServerConfig) *AuthZ {
-	rbacAuthorizer, err := NewAuthorizer()
+	rbacAuthorizer, err := rbac.NewAuthorizer()
 	if err != nil {
 		panic(err)
 	}
@@ -128,7 +124,7 @@ var restrictedWithBearerToken = []string{
 // At the moment, there's no further bearer auth because bearer tokens are admin-scoped
 // and we check that the bearer token has access to the tenant in the authn step.
 func (a *AuthZ) handleBearerAuth(c echo.Context, r *middleware.RouteInfo) error {
-	if operationIn(r.OperationID, restrictedWithBearerToken) {
+	if rbac.OperationIn(r.OperationID, restrictedWithBearerToken) {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Not authorized to perform this operation")
 	}
 
@@ -155,7 +151,7 @@ func (a *AuthZ) ensureVerifiedEmail(c echo.Context, r *middleware.RouteInfo) err
 		return nil
 	}
 
-	if operationIn(r.OperationID, permittedWithUnverifiedEmail) {
+	if rbac.OperationIn(r.OperationID, permittedWithUnverifiedEmail) {
 		return nil
 	}
 
@@ -176,108 +172,4 @@ func (a *AuthZ) authorizeTenantOperations(tenantMemberRole sqlcv1.TenantMemberRo
 	// NOTE(abelanger5): this should be default-deny, but there's not a strong use-case for restricting member
 	// operations at the moment. If there is, we should modify this logic.
 	return nil
-}
-
-func operationIn(operationId string, operationIds []string) bool {
-	for _, id := range operationIds {
-		if strings.EqualFold(operationId, id) {
-			return true
-		}
-	}
-
-	return false
-}
-
-type Authorizer struct {
-	permissionMap PermissionMap
-}
-
-func NewAuthorizer() (*Authorizer, error) {
-	permMap, err := LoadYaml()
-	if err != nil {
-		return nil, err
-	}
-	return &Authorizer{
-		permissionMap: *permMap,
-	}, nil
-}
-
-func (a *Authorizer) IsAuthorized(role sqlcv1.TenantMemberRole, operation string) bool {
-	return a.permissionMap.HasPermission(string(role), operation)
-}
-
-type Role struct {
-	Inherits    *[]string
-	Permissions *[]string
-}
-
-type PermissionError struct {
-	Message string
-}
-
-func (e *PermissionError) Error() string {
-	return e.Message
-}
-
-type PermissionMap struct {
-	Roles map[string]*Role
-}
-
-func (p *PermissionMap) HasPermission(roleName string, operation string) bool {
-	curRole := p.Roles[roleName]
-	if curRole.Permissions != nil {
-		inRole := operationIn(operation, *curRole.Permissions)
-		if inRole {
-			return true
-		}
-	}
-	if curRole.Inherits != nil {
-		for _, inheritedRoleName := range *curRole.Inherits {
-			if p.HasPermission(inheritedRoleName, operation) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func (p *PermissionMap) ValidInheritance(roleName string) error {
-	if p.Roles[roleName].Inherits == nil {
-		return nil
-	}
-	for _, inheritedRole := range *p.Roles[roleName].Inherits {
-		_, ok := p.Roles[inheritedRole]
-		if !ok {
-			return &PermissionError{
-				Message: fmt.Sprintf("%s inherits from %s which does not exist", roleName, inheritedRole),
-			}
-		}
-	}
-	return nil
-}
-
-func (p *PermissionMap) Validate() error {
-	for roleName := range p.Roles {
-		if err := p.ValidInheritance(roleName); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func LoadYaml() (*PermissionMap, error) {
-	_, yamlPath, _, _ := runtime.Caller(0)
-	yamlFile, err := os.ReadFile(filepath.Join(filepath.Dir(yamlPath), "rbac.yaml"))
-	if err != nil {
-		return nil, err
-	}
-	var yamlContents PermissionMap
-	err = yaml.Unmarshal(yamlFile, &yamlContents)
-	if err != nil {
-		return nil, err
-	}
-	if err := yamlContents.Validate(); err != nil {
-		return nil, err
-	}
-	return &yamlContents, nil
 }
