+++ api/v1/server/rbac/rbac_test.go
@@ -1,11 +1,11 @@
-package authz
+package rbac
 
 import (
 	"testing"
 
-	"github.com/hatchet-dev/hatchet/api/v1/server/middleware"
 	"github.com/hatchet-dev/hatchet/api/v1/server/oas/gen"
 	"github.com/hatchet-dev/hatchet/pkg/repository/sqlcv1"
+
 	"github.com/stretchr/testify/assert"
 )
 
@@ -22,15 +22,6 @@ var adminAndOwnerOnly = []string{
 	"ApiTokenUpdateRevoke",
 }
 
-func createRouteInfo(operationId string) *middleware.RouteInfo {
-	return &middleware.RouteInfo{
-		OperationID: operationId,
-		Security:    nil,
-		Resources:   nil,
-		Route:       nil,
-	}
-}
-
 func operationIdsFromSpec() []string {
 	spec, _ := gen.GetSwagger()
 	allOperationIds := make([]string, 0)
@@ -45,19 +36,14 @@ func operationIdsFromSpec() []string {
 func TestAuthorizeTenantOperations(t *testing.T) {
 	r, err := NewAuthorizer()
 	assert.Nil(t, err)
-	az := AuthZ{
-		config: nil,
-		rbac:   r,
-		l:      nil,
-	}
 	allOperations := operationIdsFromSpec()
 	for _, operationId := range allOperations {
-		assert.Equal(t, az.authorizeTenantOperations(sqlcv1.TenantMemberRoleADMIN, createRouteInfo(operationId)), nil)
-		assert.Equal(t, az.authorizeTenantOperations(sqlcv1.TenantMemberRoleOWNER, createRouteInfo(operationId)), nil)
-		if operationIn(operationId, adminAndOwnerOnly) {
-			assert.NotEqual(t, az.authorizeTenantOperations(sqlcv1.TenantMemberRoleMEMBER, createRouteInfo(operationId)), nil)
+		assert.Equal(t, r.IsAuthorized(sqlcv1.TenantMemberRoleADMIN, operationId), true)
+		assert.Equal(t, r.IsAuthorized(sqlcv1.TenantMemberRoleOWNER, operationId), true)
+		if OperationIn(operationId, adminAndOwnerOnly) {
+			assert.Equal(t, r.IsAuthorized(sqlcv1.TenantMemberRoleMEMBER, operationId), false)
 		} else {
-			assert.Equal(t, az.authorizeTenantOperations(sqlcv1.TenantMemberRoleMEMBER, createRouteInfo(operationId)), nil)
+			assert.Equal(t, r.IsAuthorized(sqlcv1.TenantMemberRoleMEMBER, operationId), true)
 		}
 	}
 }
